requirements.txt
aiogram~=2.25.1
kinopoisk-dev==0.2.0
requests~=2.28.2
Pillow~=9.4.0
python-dotenv~=1.0.0